% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{api_aggregation}
\alias{api_aggregation}
\title{Helper function to create an API aggregation}
\usage{
api_aggregation(aggregation_func, variable_name)
}
\arguments{
\item{aggregation_func}{Aggregation function}

\item{variable_name}{Name of variable to aggregate}
}
\value{
aggregation structure
}
\description{
Helper function to create an API aggregation
}
\examples{
api_aggregation(aggregation_func = 'sum', variable_name = 'population')
}
\keyword{query}
