% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{api_fields}
\alias{api_fields}
\title{This function structures api fields to query}
\usage{
api_fields(fields_list = list())
}
\arguments{
\item{fields_list}{A list of fields to query}
}
\value{
fields list
}
\description{
This function structures api fields to query
}
\examples{
api_fields(list('year', 'geoid2', list(population = 'state_population')))
}
\keyword{query}
