% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{api_filter}
\alias{api_filter}
\title{Helper function to create an API filter}
\usage{
api_filter(filter_variable, filter_type, filter_value)
}
\arguments{
\item{filter_variable}{Variable filtered on}

\item{filter_type}{Filter type ('eq' or 'in', e.g.)}

\item{filter_value}{Value used for filter}
}
\value{
filter structure
}
\description{
Helper function to create an API filter
}
\examples{
api_filter(filter_type = 'eq', filter_variable = 'year', filter_value = 2018)
}
\keyword{query}
