% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{api_query_params}
\alias{api_query_params}
\title{This function structures an API query to submit to the StratoDem Analytics API}
\usage{
api_query_params(table, query_type = "COUNT", data_fields, data_filters,
  aggregations = list(), groupby = list(),
  median_variable_name = NULL)
}
\arguments{
\item{table}{Table name}

\item{query_type}{Type of query - defaults to 'COUNT'}

\item{data_fields}{List of fields to return}

\item{data_filters}{List of filters for the query}

\item{aggregations}{List of aggregations for the query}

\item{groupby}{List of variables for GROUP BY}
}
\value{
query parameters
}
\description{
This function structures an API query to submit to the StratoDem Analytics API
}
\examples{
api_query_params('my-table', data_fields=list(), data_filters=list())
}
\keyword{query}
