% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{between_filter}
\alias{between_filter}
\title{Helper function to create a between filter}
\usage{
between_filter(filter_variable, filter_value)
}
\arguments{
\item{filter_variable}{Variable filtered on}

\item{filter_value}{Value used for filter}
}
\value{
filter structure
}
\description{
Helper function to create a between filter
}
\examples{
between_filter(filter_variable = 'year', filter_value = c(2013, 2018))
}
\keyword{query}
