% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{drivetime_filter}
\alias{drivetime_filter}
\title{Helper function to create a drivetime filter}
\usage{
drivetime_filter(latitude, longitude, minutes)
}
\arguments{
\item{latitude}{Latitude of target}

\item{longitude}{Longitude of target}

\item{minutes}{Minutes drivetime}
}
\value{
filter structure
}
\description{
Helper function to create a drivetime filter
}
\examples{
drivetime_filter(latitude = 42.7, longitude = -120.38, minutes = 30)
}
\keyword{query}
