% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{eq_filter}
\alias{eq_filter}
\title{Helper function to create an EQ filter}
\usage{
eq_filter(filter_variable, filter_value)
}
\arguments{
\item{filter_variable}{Variable filtered on}

\item{filter_value}{Value used for filter}
}
\value{
filter structure
}
\description{
Helper function to create an EQ filter
}
\examples{
eq_filter(filter_variable = 'year', filter_value = 2018)
}
\keyword{query}
