% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{mean_aggregation}
\alias{mean_aggregation}
\title{Helper function to create an API mean aggregation}
\usage{
mean_aggregation(variable_name)
}
\arguments{
\item{variable_name}{Name of variable to aggregate}
}
\value{
aggregation structure
}
\description{
Helper function to create an API mean aggregation
}
\examples{
mean_aggregation(variable_name = 'population')
}
\keyword{query}
