% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{mean_query_params}
\alias{mean_query_params}
\title{This function structures an API query that gets the average value of the target "mean_variable_name" to submit to the StratoDem Analytics API}
\usage{
mean_query_params(table, data_fields, data_filters, mean_variable_name,
  aggregations = list(), groupby = list(), order = list(),
  join = NA, on = NA)
}
\arguments{
\item{table}{Table name}

\item{data_fields}{List of fields to return}

\item{data_filters}{List of filters for the query}

\item{mean_variable_name}{Name of mean variable to use}

\item{aggregations}{List of aggregations for the query}

\item{groupby}{List of variables for GROUP BY}
}
\value{
query parameters
}
\description{
This function structures an API query that gets the average value of the target "mean_variable_name" to submit to the StratoDem Analytics API
}
\examples{
# Returns average household income by state in 2018
api_query_params(
  table = 'incomeforecast_state_annual_mean_household_income',
  data_fields = api_fields(fields_list = list('geoid2', list(mean_value = 'mean_hhi'))),
  data_filters = list(eq_filter(filter_variable = 'year', filter_value = 2018)),
  aggregations = list(),
  mean_variable_name = 'mean_household_income',
  groupby = c('geoid2'),
  order = c('geoid2'))
}
\keyword{query}
