% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{median_query_params}
\alias{median_query_params}
\title{This function structures an API query to submit to the StratoDem Analytics API}
\usage{
median_query_params(table, data_fields, data_filters, median_variable_name,
  aggregations = list(), groupby = list())
}
\arguments{
\item{table}{Table name}

\item{data_fields}{List of fields to return}

\item{data_filters}{List of filters for the query}

\item{median_variable_name}{Name of median variable to use}

\item{aggregations}{List of aggregations for the query}

\item{groupby}{List of variables for GROUP BY}
}
\value{
query parameters
}
\description{
This function structures an API query to submit to the StratoDem Analytics API
}
\examples{
median_query_params('my-table', data_fields=list(), data_filters=list())
}
\keyword{query}
