% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{mile_radius_filter}
\alias{mile_radius_filter}
\title{Helper function to create a mile radius filter}
\usage{
mile_radius_filter(latitude, longitude, miles)
}
\arguments{
\item{latitude}{Latitude of target}

\item{longitude}{Longitude of target}

\item{miles}{Size of radius in miles}
}
\value{
filter structure
}
\description{
Helper function to create a mile radius filter
}
\examples{
mile_radius_filter(latitude = 42.7, longitude = -120.38, miles = 5)
}
\keyword{query}
