% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{ne_filter}
\alias{ne_filter}
\title{Helper function to create a "not equal to" filter}
\usage{
ne_filter(filter_variable, filter_value)
}
\arguments{
\item{filter_variable}{Variable filtered on}

\item{filter_value}{Value used for filter}
}
\value{
filter structure
}
\description{
Helper function to create a "not equal to" filter
}
\examples{
ne_filter(filter_variable = 'year', filter_value = 2018)
}
\keyword{query}
