% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_query.R
\name{submit_api_query}
\alias{submit_api_query}
\title{This function submits an API query object as JSON to the StratoDem Analytics API}
\usage{
submit_api_query(query, apiToken)
}
\arguments{
\item{query}{Query object to submit to the StratoDem Analytics API}

\item{apiToken}{API token for authentication}
}
\value{
data.frame
}
\description{
This function submits an API query object as JSON to the StratoDem Analytics API
}
\examples{
submit_api_query(list(), 'my-api-token')
}
\keyword{query}
