% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{sum_aggregation}
\alias{sum_aggregation}
\title{Helper function to create an API sum aggregation}
\usage{
sum_aggregation(variable_name)
}
\arguments{
\item{variable_name}{Name of variable to aggregate}
}
\value{
aggregation structure
}
\description{
Helper function to create an API sum aggregation
}
\examples{
sum_aggregation(variable_name = 'population')
}
\keyword{query}
